Rename things, add ensmallen_graph call
